Create YouTubeExpiredOverlay component
